Fix the mouse inside the macOS build.
When we don't have the editor on, C4WindowController.viewport is nil.
In that case, we have to use the ModePlay mouse handling, not the
ModeEdit handling.  Change C4OpenGLView.mouseEvent to handle this.

Also, when compiled with WITH_QT_EDITOR=1, we need to check whether the
requested window type is a W_Viewport before assuming that we're embedding
it inside the editor.  This handles the case where WITH_QT_EDITOR=1,
but the app is running without the --editor flag on the command line.

Both these changes make the macOS event logic more like
C4ConsoleQtViewportView / C4WindowSDL and less like C4WindowWin32.
The Win32 platform code is different because it uses a different WndProc
entirely for W_Viewport vs W_Fullscreen, but the other platforms don't do
that.

Also note that W_Fullscreen doesn't mean a full-screen window (!)  It
means "not inside the editor", which on macOS can either be a framed
window or full-screen mode.

With these changes, the macOS build of the game now works properly in
non-editor mode.  It seems to "work" in editor mode, in that everything
seems to be there, but the game embedded in the editor is unusably slow.
I don't know why.